+++ repos/ktor/common/src/commonMain/kotlin/dev/inmo/micro_utils/repos/ktor/common/one_to_many/OneToManyParametersNames.kt
@@ -1,8 +1,0 @@
-package dev.inmo.micro_utils.repos.ktor.common.one_to_many
-
-@Deprecated("Replaced", ReplaceWith("keyParameterName", "dev.inmo.micro_utils.repos.ktor.common.keyParameterName"))
-const val keyParameterName = dev.inmo.micro_utils.repos.ktor.common.keyParameterName
-@Deprecated("Replaced", ReplaceWith("valueParameterName", "dev.inmo.micro_utils.repos.ktor.common.valueParameterName"))
-const val valueParameterName = dev.inmo.micro_utils.repos.ktor.common.valueParameterName
-@Deprecated("Replaced", ReplaceWith("reversedParameterName", "dev.inmo.micro_utils.repos.ktor.common.reversedParameterName"))
-const val reversedParameterName = dev.inmo.micro_utils.repos.ktor.common.reversedParameterName